Increase the number of possible tiles to 2^32
There's no reason to be memory thrifty here. It's possible to programmatically
inflate the size of an input tile set with flips and mirror operations.

+++ crawl-ref/source/domino.cc
@@ -49,7 +49,7 @@ Adjacency::Adjacency()
     for (size_t i = FIRST_DIRECTION; i <= LAST_DIRECTION; ++i)
     {
         Direction d = static_cast<Direction>(i);
-        permitted_[d] = new set<uint8_t>();
+        permitted_[d] = new set<uint32_t>();
     }
 }
 
@@ -62,18 +62,18 @@ Adjacency::~Adjacency()
     }
 }
 
-bool Adjacency::adjacent(Direction d, set<uint8_t>& open)
+bool Adjacency::adjacent(Direction d, set<uint32_t>& open)
 {
     intersection(open, *permitted_[d]);
     return !open.empty();
 }
 
-bool Adjacency::permitted(Direction d, uint8_t id)
+bool Adjacency::permitted(Direction d, uint32_t id)
 {
     return *permitted_[d]->find(id) != *permitted_[d]->end();
 }
 
-void Adjacency::add(uint8_t adjacency, const set<Direction>& dir)
+void Adjacency::add(uint32_t adjacency, const set<Direction>& dir)
 {
     set<Direction>::iterator itr = dir.begin();
     for (; itr != dir.end(); ++itr)
